+++ R/HDcox_splsdacox_mixomics.R
@@ -293,6 +293,7 @@ splsdacox_mixOmics <- function (X, Y,
 #' @param remove_near_zero_variance Logical. If remove_near_zero_variance = TRUE, remove_near_zero_variance variables will be removed.
 #' @param remove_zero_variance Logical. If remove_zero_variance = TRUE, remove_zero_variance variables will be removed.
 #' @param toKeep.zv Character vector. Name of variables in X to not be deleted by (near) zero variance filtering.
+#' @param remove_variance_at_fold_level Logical. Remove variance at fold level (T) or before split the data (F-default).
 #' @param remove_non_significant_models Logical. If remove_non_significant_models = TRUE, non-significant models are removed before computing the evaluation.
 #' @param remove_non_significant Logical. If remove_non_significant = TRUE, non-significant variables in final cox model will be removed until all variables are significant (forward selection).
 #' @param alpha Numeric. Cutoff for establish significant variables. Below the number are considered as significant (default: 0.05).
@@ -328,7 +329,7 @@ cv.splsdacox_mixOmics <- function(X, Y,
                         vector = NULL,
                         x.center = TRUE, x.scale = FALSE,
                         y.center = FALSE, y.scale = FALSE,
-                        remove_near_zero_variance = T, remove_zero_variance = T, toKeep.zv = NULL,
+                        remove_near_zero_variance = T, remove_zero_variance = T, toKeep.zv = NULL, remove_variance_at_fold_level = F,
                         remove_non_significant_models = F, remove_non_significant = F, alpha = 0.05,
                         MIN_NVAR = 10, MAX_NVAR = 1000, n.cut_points = 5,
                         MIN_AUC_INCREASE = 0.01,
@@ -364,13 +365,17 @@ cv.splsdacox_mixOmics <- function(X, Y,
   max.ncomp <- check.maxPredictors(X, Y, MIN_EPV, max.ncomp, verbose = verbose)
 
   #### REQUIREMENTS
-  lst_dnz <- deleteZeroOrNearZeroVariance(X = X,
-                                          remove_near_zero_variance = remove_near_zero_variance,
-                                          remove_zero_variance = remove_zero_variance,
-                                          toKeep.zv = toKeep.zv,
-                                          freqCut = 95/5)
-  X <- lst_dnz$X
-  variablesDeleted <- lst_dnz$variablesDeleted
+  if(!remove_variance_at_fold_level & (remove_near_zero_variance | remove_zero_variance)){
+    lst_dnz <- deleteZeroOrNearZeroVariance(X = X,
+                                            remove_near_zero_variance = remove_near_zero_variance,
+                                            remove_zero_variance = remove_zero_variance,
+                                            toKeep.zv = toKeep.zv,
+                                            freqCut = 95/5)
+    X <- lst_dnz$X
+    variablesDeleted <- lst_dnz$variablesDeleted
+  }else{
+    variablesDeleted <- NULL
+  }
 
   #### #
   # CV #
@@ -397,7 +402,7 @@ cv.splsdacox_mixOmics <- function(X, Y,
                                        MIN_AUC_INCREASE = MIN_AUC_INCREASE,
                                        EVAL_METHOD = EVAL_METHOD,
                                        x.center = x.center, x.scale = x.scale, y.center = y.center, y.scale = y.scale,
-                                       remove_near_zero_variance = F, remove_zero_variance = F, toKeep.zv = NULL,
+                                       remove_near_zero_variance = remove_variance_at_fold_level, remove_zero_variance = F, toKeep.zv = NULL,
                                        remove_non_significant = remove_non_significant,
                                        total_models = total_models, max.iter = max.iter, tol = tol, PARALLEL = PARALLEL, verbose = verbose)
 
@@ -451,6 +456,12 @@ cv.splsdacox_mixOmics <- function(X, Y,
   if(w_AUC!=0){
     #total_models <- ifelse(!fast_mode, n_run * max.ncomp, k_folds * n_run * max.ncomp)#inside get_COX_evaluation_AUC
 
+    #times should be the same for all folds
+    #calculate time vector if still NULL
+    if(is.null(times)){
+      times <- getTimesVector(Y)
+    }
+
     lst_df <- get_COX_evaluation_AUC(comp_model_lst = comp_model_lst,
                                      lst_X_test = lst_X_test, lst_Y_test = lst_Y_test,
                                      df_results_evals = df_results_evals, times = times,
@@ -458,7 +469,7 @@ cv.splsdacox_mixOmics <- function(X, Y,
                                      max.ncomp = max.ncomp, n_run = n_run, k_folds = k_folds,
                                      MIN_AUC_INCREASE = MIN_AUC_INCREASE, MIN_AUC = MIN_AUC, MIN_COMP_TO_CHECK = MIN_COMP_TO_CHECK,
                                      w_AUC = w_AUC, #total_models = total_models,
-                                     method.train = pkg.env$splsdacox_mixomics, PARALLEL = F)
+                                     method.train = pkg.env$splsdacox_mixomics, PARALLEL = F, verbose = verbose)
 
     df_results_evals_comp <- lst_df$df_results_evals_comp
     df_results_evals_run <- lst_df$df_results_evals_run
